moved snackbar component from toolbar-folder into own folder
