+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Snackbar/Snackbar.js
