+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Snackbar/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Snackbar from './Snackbar';
+
+export default Snackbar;
