+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Snackbar/tests/Snackbar.test.js
