+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Toolbar/tests/Toolbar.test.js
@@ -3,7 +3,7 @@ import {render} from 'enzyme';
 import React from 'react';
 import Button from '../Button';
 import Controls from '../Controls';
-import Snackbar from '../Snackbar';
+import Snackbar from '../../Snackbar';
 import Toolbar from '../Toolbar';
 
 jest.mock('../../../utils/Translator', () => ({
